requirements/base.txt
pytz==2024.1  # https://github.com/stub42/pytz
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.3  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
gunicorn==23.0.0  # https://github.com/benoitc/gunicorn
uvicorn[standard]==0.24.0.post1
channels
channels_redis

# Django
# ------------------------------------------------------------------------------
django==4.2.20  # LTS https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.4  # https://github.com/django-compressor/django-compressor
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-markdownify==0.9.3  # https://django-markdownify.readthedocs.io/en/latest/install_and_usage.html
django-ipware==7.0.1  # https://github.com/un33k/django-ipware
# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3 # Included explicitly since it's defined optional by djangorestframework
setuptools # As coreapi needs 'pkg_resources' to load and 'pkg_resources' as been migrated to 'setuptools' since python3.12
django-cors-headers==4.3.1 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
django-q2==1.6.2
django-minio-storage==0.5.7
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil==2.9.0.post0
djangorestframework_simplejwt==5.3.1
python-magic==0.4.27
dj-rest-auth==6.0.0
pynacl==1.5.0
django-allauth-2fa==0.11.1
# django-serverless-cron==0.1.3 # https://github.com/paulonteri/django-serverless-cron/tree/master/django_serverless_cron
django-serverless-cron @ git+https://github.com/TontonSancho/django-serverless-cron.git

# Analysis
communityid==1.5.0
yara-python==4.5.0  # https://github.com/VirusTotal/yara-python
pcapng-utils>=1.0.9  # https://github.com/PiRogueToolSuite/pcapng-utils

# Mandolin
mandolin-python-client>=1.0.2

# Colander
colander-data-converter==1.0.7
